requirements.txt
pyyaml==6.0.2
pandas==2.2.2
python-dotenv==1.0.1
langchain-openai==0.2.14
langgraph==0.2.60
langgraph-checkpoint-sqlite==2.0.1
langchain_community==0.3.13
pyaudio==0.2.14
faster_whisper==1.1.0
langchain==0.3.13
beautifulsoup4==4.12.3
pyttsx3==2.90
torch==2.5.1
torchaudio==2.5.1